Fixed Gregorian bug
